build: enable -Werror gate on Linux Release; clean baseline warnings (#23)
Closes #16. Drives the project warning baseline added in #16 to zero
across all targets and flips ECONLIFE_WARN_AS_ERROR=ON for the
ubuntu-latest x Release CI cell so regressions are merge-blocking.

Highlights:

* Apply econlife_set_warnings / econlife_maybe_werror to the four
  test/benchmark CMakeLists that were silently bypassing the gate.
* Drop named parameters in ITickModule's default no-op virtuals
  (single largest source of -Wunused-parameter at ~200 hits).
* Pass SYSTEM to every FetchContent_Declare so dep headers stop
  surfacing through the project warning baseline.
* fix(real_estate): wire avg_property_value through Province
  copy ctor + RegionDelta + apply_deltas, route via region_id
  rather than province_idx (codex P1/P2).
* fix(drug_economy): thread per-tier output_quality into
  BusinessDelta.output_quality_update — the spec called for it
  and the module was discarding the computed value.
* fix(world_generator): initialise soil_multiplier to 1.0f
  (-Wmaybe-uninitialized was a real flow-analysis miss).
* fix(settlement_generator): bail when owner_pool is empty
  rather than feeding 0 to next_uint() and indexing UB.
* Mechanical cleanup of unused-parameter / unused-variable /
  sign-conversion / double-promotion / missing-field-initializers
  / unused-function across modules and tests; scaffolding
  helpers preserved with [[maybe_unused]].

1313 ctest cases pass under -Werror Release locally; sanitizer
matrix and cross-platform determinism gate green on PR #23.

+++ simulation/core/world_gen/technology_catalog.cpp
@@ -127,7 +127,7 @@ bool TechnologyCatalog::load_ceilings_csv(const std::string& filepath) {
 
         MaturationCeilingEntry entry;
         entry.node_key = fields[0];
-        for (int i = 0; i < 10 && i + 1 < static_cast<int>(fields.size()); ++i) {
+        for (size_t i = 0; i < 10 && i + 1 < fields.size(); ++i) {
             entry.era_ceilings[i] = std::strtof(fields[i + 1].c_str(), nullptr);
         }
